[NCCL][Blocking Wait] Log set exceptions when checking for exceptions in blocking wait
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54558

In blocking wait's polling synchronization loop, we frequently call checkAndSetException() as part of isCompleted() to check the status of nccl operations. It would be useful to log here in case we encounter any exceptions (which are later thrown by `checkAndThrowException`).

Also slightly refactors code previously added to make use of a helper function to get the error message given an `std::exception_ptr`.
ghstack-source-id: 124942366

Differential Revision: [D27136202](https://our.internmc.facebook.com/intern/diff/D27136202/)