torch.testing allclose -> close
Right now the functions have divergent names with one postfixed `_equal` and the other `_allclose`. I've opted to use `_(equal|close)` over `_all(equal|close)` think it is a reasonable assumption that all values need to be equal or close for this pass even without explicitly naming the function this way.

ghstack-source-id: 0db0b3ff736620caf9b2d974c90e45e3c86f32b4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54781